Resolution and new enemy type
I have the UI shrinking and stretching correctly as long as we are
locked to 4:3. There is an issue specifically with 1600x1200 resolution
and I do not know why.
A new enemy type has been made. It has it's own sprite and AI. A spawn
wave that features them is now in the level.