Configurable AI
Experimenting with making the AI routines and the spawners for those
routines more configurable to get more mileage out of them.
The Tank enemy (Enemy_G) can now point in any of the four cardinal
directions. I made a new spawner to take advantage of that fact so it
has tanks coming up and down from both ends of the screen.
It's something to think about in future development that might prove
useful.