Fixed infinite scrolling
Repositioning after an infinite background piece falls offsecreen is now
greatly improved. The math before was incorrect and relied on the size
of the window itself being consistent and the background being a
particular size. Now as long as a background covers the entire camera
view screen or more, it should just work.