Powerups
There is now facility for enemies to drop powerups when destroyed. All
the powerups themselves still have a long way to go, but the speed
powerup works and an enemy can be assigned to drop one. I want to
eventually have them trigger to drop when a group of enemies is
destroyed.