PowerupGroups
Now powerups can spawn when a group of enemies are destroyed and not a
solitary one. I learned an important Unity lesson. On a MonoBehaviour,
Start() is not called when Instantiate is. Start gets called before the
first OnUpdate(). That may explain my frustration I was having before
with class members "resetting". Awake() gets called on instantiation.
EnemyAI4 has had all of it's Start() code transferred to Awake().