Bug fix and tweaks
I tweaked some ai and enemy spawn placements. A bug existed with
multiple powerup groups in the same place. When powerup ID's needed to
be decremented it wouldn't do it for the individual enemies, only the
group object. A fix has been implemented. Player death has a new
animation but there are problems with it not transitioning to the
flashing animation when the ship is coming back onscreen.