Powerup movement finished
The arcing floaty movement of a powerup object is achieved by Slerping
across two vectors. Each update loop have the last slerped vector saved
and subtract it from the new to be slerped vector. That difference now
becomes the objects velocity. Then store the new slerped vector over the
old one. Over time this can create nice arcs if you use vector pairs
like ((1,1) (-1,1)) or ((0,1), (-1, 0)). If you preserve the direction
of the velocity vector but change it's magnitude you can change the
speed of the movement.