Final enemy adjustments for some time
This is the last gameplay tweak commit until maybe the very end.
Tomorrow starts work on improving audio and sound effects.
The tank enemy got its own solo spawn script and it is used to fill some
dead space with something to shoot at.
I am not committing to this being the last time I change enemy placement
before launch or anything like that, but I am putting it down to work on
audio according to schedule.