Sound Effects
Sound effects for the boss being destroyed, the player getting shot, and
the boss charging at the player have been made. Code has been changed in
the level finished and gameover code to give time for sounds to play out
before game over or stage clear signs are shown and the level is ended.