Pausing
A new interface PauseableItem allows any implementing class to be paused
by the GameController. The GameController is using an observer pattern,
Pauseable items are responsible for registering and unregistering
themselves to the GameController as well as implementing the interface.
I am not done implementing, cleaning, and documenting everything but the
functionality is in and it works.