Pause Menu
I hacked together a pause menu that brings up text that says "Title?"
and below it are "yes" and "no". The same selector used in the main menu
is used here as well. If the player selects no, the game just unpauses
and resumes. If the pick yes the main menu is loaded.
It was done hastily. I want to Integrat the IngameMenu class into
UIController somehow because it should be. LIke it should be implemented
it terms of it. Also I managed to have UI elements appear without having
them be there all the time and just increase opacity. Maybe I should
apply that to the other UI elements where I set default opacity to 0 and
when I "show" them, I just increase that opacity. No point having those
objects hang around and loading and instantiating them doesn't take that
long, they are single objects.