New player ship sprite
The new ship sprite has an up and down movement sprite that it animates
into. When hit and invincible there are separate animations for moving
up or down while invincible too. All the transitions work, but the code
implementation is rushed and should get a second pass. The player now
has 7 animation states to juggle.