Enemy pattern tweaks
I added two tanks to the beginning of section 2. There was some dead air
there that I wanted to fill with something.  They go up and down towards
each other and will be seen spawning at the right corners of the screen.
EnemySpawn15 where 10 ships do that diagonal thing has been reworked.
The ai now shoots at set times which let me make a better looking
pattern. They all shoot at the start and midle of their path, with the
beginning shots waiting longer and longer for the later ships. This
forces players close to the middle stream of bullets in a cone.