Fullscreen resolution. PowerupGroup fixes. Sprite animation tweaks
The game now forces windowed. If the player selects full screen the main
menu script will force it back down to a window of the selected
resolution.

There were still issues with powerup groups. This time being that if a
group was removed there could still be enemies from that group existing
and still having that ID. This mostly happened only when those enemies
were destroyed close to or at the same time. This caused issues with
additional groups having their ID's shifted down twice instead of once,
causing that group to not behave or spawn it's powerup. Now when a group
is removed the first thing it does is set all other squadmates ID's to
-1 and then shifts everyone else around. This prevents that squad
removal from happening more times than intended.

I changed and added detail to the turret and diamond enemy. They look
more congruent with the rest of the game's style.