Fixes to powerup objects and speed marker placement
There were issues with powerup objects not behaving properly when not
picked up. They were not destroyed when they left the screen so they
would be offscreen forever. It was also not pauseable. If the player
paused while one was on screen it would still move around. I fixed that,
stopping velocity on pause and returning to previous velocity when
unpaused.

The placement of the speed level on the UI was not dependant on screen
pixel width, making it appear differently spaced depending on what
resolution the game was running at. The offset of the X value to
additional speed markers are not a function of the screen width:
Screen.width/35f.

I redid the tank sprite. It matches style more closely with the other
enemies now.