Changed controls and removed boss pattern
Tested the Linux build and the controls had an issue withm ovement
controls. using the arrow keys, the veritcal and horizontal axis would
not reset so the character would be moving into the corners. I have
changed the controls to work differently 4 different button presses and
not 2 axis.
I took the charge attack pattern away from the boss because it's bad.