GameController testing
I painstakingly made 2 tests that are dependant on a specific
environment and even the order in which they are executed between each
other. They are not good tests. In order to write better tests, I need
to refactor code. I have started a rewrite of GameController. The
MonoBehaviour is one module, the logic it  interacts with is another.
Because the GameController has so many responsibilities, it will be
split up into other modules, each of which will be an interface with a
real implementation that can potentially have test doubles along side it
if need be.