Additional work on GameControllerRewrite
Added 5 new methods to the IPowerupGroupController. These cover the the
duties of Adding and removing squads, getting the next squad ID, getting
the number of squads currently in play, and checking if a squad is
currently in the list. Checking if a squad is there can be done by
passing the PowerupGroup object itself or just the squad's ID as an
integer.
I took SpawnPowerupAtPosition out of the interface. That is heavily
reliant on Unity API so I determined it should be part of the
MonoBehaviour. I implemented that in GameControllerBehaviour.
I wrote implementations for all of the function added to the interface
definition in the implementation class PowerupGroupController. I also
added fields for next squad ID and the squad list. I did not make these
part of the interface as properties. Another implementation could use a
HashMap instead of a list if it wanted to. And keeping track of squad
ID's could be done other ways as well for some other implementation.
Also good for mocking to keep these out of the interface.