Decoupling PowerupGroup from GameObject
PowerupGroup used to load a GameObject that would be the powerup it
would spawn once it was activated. That GameObject would then get passed
to GameController to be positioned and instantiated.
Now instead of doing that it saves a Powerup type enum and passes that
to the GameController instead. Now it is the responsibility of the
GameController to load the GameObject and spawn it.
The PowerupGroup functions ReturnPowerupObject and SetPowerupObject have
been removed and all references to them have been replaced with the
equivalent ReturnPowerupType and SetPowerupType.
This is one step in refacotrin the IPowerupGroupController to be totally
decoupled from Unity API