Removed InitializePowerupGroup from IPowerupControllerInterface
I decided this was not a necessary part of the interface. Anything that
should be done in a funciton like this should probably just get done in
the constructor.
I removed a commented out Debug.LogError in GameController because it
was old and needed to be removed a long time ago.