Beginning a second level
The level is water themed and I have a rough background. It is
selectable from the main menu as well. Instead of "Start" the title
screen has "Level 1" and "Level 2".
Basically nothing is done on the level and there isn't a way to exit it
other than opening the menu and quitting. Intend to design this level
from the ground up with the shield mechanic while tweaking the shield
mechanic to work better with the level.