Major shield appearance and code update
The shield has had a major rework. The shield functionality has now been
split up into three components for testability following the humble
object pattern. That would be the IPlayerShield interface, PlayerShield
implementation, and PlayerShieldBehaviour MonoBehaviour. The Player
class now hold reference to an interface and not a GameObject, major
improvment. This is all in line with the changes being made in the
TestFramework branch.

The other part of this update is the look of the shield. It has a
sprite, a gameobject, and a prefab that instantiates it. That prefab has
animations with animation states implemented. The animations show 3
different levels of shield energy based on color.