First integration of rewrite
I have integrated all of the new rewritten GameController code into the
the rest of the code. The old GameController code is still there beside
it for the most part, I will have to comment all of that out while I
test out and fix the new implementation. Some of the interfaces like
IPowerupGroupController were modified as new requirements came up.