Work on pause controller, level content including boss of level 3, and fixed bullets not appearing on creation bug
It used to be that during the pause state, if a coroutine inside another
coroutine was stopped and it was in the middle of a WaitForSeconds, the
timer would keep going. This caused issue when pausing during the
player's kill sequence. THere are new static helper functions in the
PauseController to let coroutines pause and pause their wait timers they
yield on.
It appeared before that newly generated bullets would not be shot. They
were actually there, it's just that bullets by default are not being
rendererd. The renderer must bet turned on before being returned to the
enemy to shoot.
More content for the first four levels. I want to finish the boss for
level 3 and then polish those first three levels up for a demo that is
half the game.